copy changes to brgy template

+++ barangay_template/admin/officials/officials-table.php
@@ -4,11 +4,6 @@ include '../../includes/dbh.inc.php';
 include '../../includes/session.inc.php';
 include '../../includes/deactivated.inc.php';
 
-// $stmt = $pdo->prepare("SELECT * FROM resident WHERE barangay_id = :barangay_id");
-// $stmt->bindParam(':barangay_id', $barangayId, PDO::PARAM_INT);
-// $stmt->execute();
-// $resident = $stmt->fetchAll(PDO::FETCH_ASSOC);
-
 // Define the SQL query to join the tables
 $stmt = $pdo->prepare("SELECT * FROM resident 
                     INNER JOIN officials ON resident.resident_id = officials.resident_id
@@ -95,37 +90,46 @@ $results = $stmt->fetchAll();
                 </div>
             </div>
 
-            <!-- Page body -->
-            <div class="page-body">
-                <table id="officials-table" class="row-border hover">
-                    <thead>
-                        <tr>
-                            <th>Full Name</th>
-                            <th>Gender</th>
-                            <th>Position</th>
-                            <th>Date Appointed</th>
-                            <th>Date End</th>
-                            <th>Action</th>
-                        </tr>
-                    </thead>
-                    <tbody>
-                        <?php foreach ($results as $resident) { ?>
+            <?php if ($logged_resident['position'] == 'Barangay Secretary') : ?>
+                <!-- Page body -->
+                <div class="page-body">
+                    <table id="officials-table" class="row-border hover">
+                        <thead>
                             <tr>
-                                <td><?php echo "$resident[firstname] $resident[middlename] $resident[lastname]" ?></td>
-                                <td><?php echo $resident['sex'] ?></td>
-                                <td><?php echo $resident['position'] ?></td>
-                                <td><?php echo date('F j, Y', strtotime($resident['date_start'])) ?></td>
-                                <td><?php echo date('F j, Y', strtotime($resident['date_end'])) ?></td>
-                                <td>
-                                    <button type="button" id="deleteBtn" class="focus:outline-none text-white bg-red-700 hover:bg-red-800 focus:ring-4 focus:ring-red-300 font-medium rounded-lg text-sm px-5 py-2.5 mr-2 mb-2 dark:bg-red-600 dark:hover:bg-red-700 dark:focus:ring-red-900">
-                                        <a href="includes/delete-officials.inc.php?id=<?php echo $resident['resident_id'] ?>">Delete</a>
-                                    </button>
-                                </td>
+                                <th>Full Name</th>
+                                <th>Gender</th>
+                                <th>Position</th>
+                                <th>Date Appointed</th>
+                                <th>Date End</th>
+                                <th>Action</th>
                             </tr>
-                        <?php } ?>
-                    </tbody>
-                </table>
-            </div>
+                        </thead>
+                        <tbody>
+                            <?php foreach ($results as $resident) { ?>
+                                <tr>
+                                    <td><?php echo "$resident[firstname] $resident[middlename] $resident[lastname]" ?></td>
+                                    <td><?php echo $resident['sex'] ?></td>
+                                    <td><?php echo $resident['position'] ?></td>
+                                    <td><?php echo date('F j, Y', strtotime($resident['date_start'])) ?></td>
+                                    <td><?php echo date('F j, Y', strtotime($resident['date_end'])) ?></td>
+                                    <td>
+                                        <?php if ($resident['position'] !== 'Barangay Secretary') : ?>
+                                            <button type="button" id="deleteBtn" class="focus:outline-none text-white bg-red-700 hover:bg-red-800 focus:ring-4 focus:ring-red-300 font-medium rounded-lg text-sm px-5 py-2.5 mr-2 mb-2 dark:bg-red-600 dark:hover:bg-red-700 dark:focus:ring-red-900">
+                                                <a href="includes/delete-officials.inc.php?id=<?php echo $resident['resident_id'] ?>">Delete</a>
+                                            </button>
+                                        <?php endif; ?>
+                                    </td>
+                                </tr>
+                            <?php } ?>
+                        </tbody>
+                    </table>
+                </div>
+            <?php else : ?>
+                <div class="page-body" style="display:flex; align-items:center; justify-content:center;">
+                    <h1>You do not have sufficient privileges to access this page.</h1>
+                </div>
+
+            <?php endif ?>
 
         </div>
     </main>
